+++ plugins/guests/suse/plugin.rb
@@ -16,6 +16,11 @@ module VagrantPlugins
         Cap::ChangeHostName
       end
 
+      guest_capability("suse", "halt") do
+        require_relative "cap/halt"
+        Cap::Halt
+      end
+
       guest_capability("suse", "configure_networks") do
         require_relative "cap/configure_networks"
         Cap::ConfigureNetworks