+++ CHANGELOG.md
@@ -15,6 +15,7 @@ BUG FIXES:
       crash Vagrant. [GH-3364]
   - commands/status: Machine readable output contains the target. [GH-3218]
   - guests/debian: Changing host name works properly. [GH-3283]
+  - guests/suse: Shutdown works correctly on SLES [GH-2775]
   - providers/virtualbox: Add missing translation for stopping status. [GH-3368]
   - providers/virtualbox: Host-only networks set cableconnected property
       to "yes" [GH-3365]